(ev_triggercheck) Decompile to the end

+++ src/event.c
@@ -12,6 +12,7 @@
 #include "scene.h"
 #include "ekrbattle.h"
 #include "bmlib.h"
+#include "ev_triggercheck.h"
 #include "event.h"
 
 // TODO: move those where they belong when possible
@@ -547,7 +548,7 @@ void SetEventTriggerState(u16 triggerId, bool8 value) {
         SetEventId(triggerId);
 }
 
-int GetEventTriggerState(u16 triggerId) {
+s8 GetEventTriggerState(u16 triggerId) {
     if (!CheckEventId(triggerId))
         return FALSE;
     return TRUE;
